reorganize UploadProgress
