+++ src/components/pages/gallery/Upload.tsx
@@ -3,7 +3,7 @@ import React, { useContext, useEffect, useRef, useState } from 'react';
 
 import { syncCollections, createAlbum } from 'services/collectionService';
 import constants from 'utils/strings/constants';
-import UploadProgress from './UploadProgress';
+import UploadProgress from '../../UploadProgress';
 
 import UploadStrategyChoiceModal from './UploadStrategyChoiceModal';
 import { SetCollectionNamerAttributes } from '../../Collections/CollectionNamer';
